refactor: Create a 'ToDontList' component
